requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Login==0.6.3
gotrue==2.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
postgrest==0.16.0
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
realtime==1.0.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
storage3==0.7.0
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
typing_extensions==4.10.0
urllib3==2.2.1
websockets==11.0.3
Werkzeug==3.0.1
Flask-CORS==3.0.10
